chore: rename hittable directory to hittables
