+++ src/render/hittable/hittable_list.c
@@ -1,4 +1,4 @@
-#include "hittable/array.h"
+#include "hittables/array.h"
 #include <float.h>
 #include <stdlib.h>
 #include <libft.h>
